+++ common/src/main/kotlin/net/spaceeye/vmod/toolgun/modes/ExtendableToolgunMode.kt
@@ -1,8 +1,10 @@
 package net.spaceeye.vmod.toolgun.modes
 
 import dev.architectury.event.EventResult
+import dev.architectury.utils.Env
 import gg.essential.elementa.components.UIContainer
 import net.minecraft.network.FriendlyByteBuf
+import net.spaceeye.vmod.reflectable.ReflectableItemDelegate
 import net.spaceeye.vmod.reflectable.SubReflectable
 import net.spaceeye.vmod.toolgun.ToolgunInstance
 import org.jetbrains.annotations.ApiStatus.NonExtendable
@@ -42,6 +44,13 @@ abstract class ExtendableToolgunMode: BaseMode, EBase, EGUIBuilder, EHUDBuilder,
     @SubReflectable val linearExtensions = mutableListOf<ToolgunModeExtension>()
     private val _extensions = mutableSetOf<ToolgunModeExtension>()
     val extensions: Collection<ToolgunModeExtension> get() = _extensions
+    lateinit var envType: BaseNetworking.EnvType
+
+    fun <T: Any> ReflectableItemDelegate<T>.onClientChange(fn: (T) -> Unit): ReflectableItemDelegate<T> =
+        this.setSetWrapper { old, new -> if (envType == BaseNetworking.EnvType.Client) { fn(new) }; new }
+
+    fun <T: Any> ReflectableItemDelegate<T>.onServerChange(fn: (T) -> Unit): ReflectableItemDelegate<T> =
+        this.setSetWrapper { old, new -> if (envType == BaseNetworking.EnvType.Server) { fn(new) }; new }
 
     fun <T: ExtendableToolgunMode> addExtensionFn(fn: (T) -> ToolgunModeExtension): T {
         val ext = fn(this as T)
@@ -144,6 +153,7 @@ abstract class ExtendableToolgunMode: BaseMode, EBase, EGUIBuilder, EHUDBuilder,
 
     final override fun init(type: BaseNetworking.EnvType) {
         if (wasInitialized) return
+        envType = type
 
         _extensions.forEach { it.preInit(this, type) }
         super.eInit(type)